Add TaskList and pause storing/loading of data.
